Fix type warning

+++ src/appvicis.c
@@ -22,7 +22,7 @@ void initVicScrMode(screen *scr, uchar bgCol, uchar fgCol, uchar chrCol) {
 	/* Clear screen to spaces */
 	(scr->clearScr)(scr, 32);
 	/* Only copy VDC character set if RAM used */
-	if (scr->chrMem != 0x1000 && scr->chrMem != 0x1800) {
+	if ((ushort) scr->chrMem != 0x1000 && (ushort) scr->chrMem != 0x1800) {
 		/* Copy VDC alt char set to VIC mem */
 		copyVdcChrMem(scr->chrMem, 0x3000, 256);
 	}
